output/framework_shorts/requirements.txt
# Knowledge Shorts Production Pipeline - Dependencies

# Core
pillow>=10.0.0          # Image generation and manipulation
pydub>=0.25.1           # Audio processing
python-dotenv>=1.0.0    # Environment variable management

# Script Generation & Templates
jinja2>=3.1.2           # Template engine for script generation
pyyaml>=6.0             # YAML configuration parsing
jsonschema>=4.17.0      # JSON schema validation

# AI & Text-to-Speech
openai>=1.0.0           # OpenAI TTS API
# elevenlabs>=0.2.0     # Alternative TTS (optional)

# Video Compilation
moviepy>=1.0.3          # Video editing wrapper for FFmpeg
srt>=3.5.0              # Subtitle/SRT file generation
numpy>=1.24.0           # Required by moviepy

# CLI & Utilities
click>=8.1.0            # CLI framework
tqdm>=4.65.0            # Progress bars
colorama>=0.4.6         # Colored terminal output
requests>=2.31.0        # HTTP requests

# Development
pytest>=7.4.0           # Testing framework
black>=23.0.0           # Code formatting
